Add/modify columns for austroasiatic languages(Japanese)
For the Japanese language, the Part of Speech seems to be blank. so I replaced it with the Normalized String which seems to be the Chinese reading of kanji, or the Romanized pronunciation of kana.

Other Austroasiatic languages, such as Chinese might have part of speech or will eventually have part of speech, so perhaps an additional columns would be useful.